backend/requirements.txt
Pillow
hachoir
exifread
geopy
pyinstaller
